requirements.txt
PySide6-Essentials==6.8.3
keyring
pyinstaller
